Add files via upload
Has basic stuff for the plants themselves, including an example prefab. More prefabs can be made relatively easily, but I may want to hold off until other stuff can be combined. Once we get inventory stuff I can add the plants drops and how it adds it to the inventory / drops it in the game (to be picked up). Once we get the player, I can add stuff like clicking on the plant to actually interact with them (harvest them or whatever). Once we get a game manager, I can make it so the plants update together and grow at the same rate (or however we want to do it).