Add files via upload
Changed a lot of scripts. Main changes include: Ability to harvest, ability to plant, proper player sprite movement, and more (there was a lot, I've forgotten most of the changes at this point). This is the current version of the game and what was in the first submission for the first playtest assignment.