Add files via upload
Lotsa stuff here. Basically added a couple of other plants (glass and titanium) that work in basically the same way as fuel (but with different values, of course). I also have a basic version of a worker bot (one of the AI things I'm workin on), though its very buggy and needs a bunch of work still.