2P final setup
Fixed the 2 player scene and adjusted the controls so it is *mostly* playable. Still can't use the inventory bar for both players and opening the inventory menu had to be disabled for the control scheme. Still need to add pests and tools. Something also seems to be wrong with planting while in the split screen.
FIRST SUBMITTED VERSION FOR PLAYTEST 2.