PestBot
I tried to do this on an old copy of the game and it worked but now its showing up that the PlantType cannot be found. Not sure what the issue is here.

In addtion,
this commit should now have the flee and the attackplants ai, wander is still in progress

wanted to commit it to the project in case someone wanted to look it over.

Also in terms of the NavMesh, there is a way to import the folder directly into Unity Project, however im not sure if you guys would have done that so if committing the NAVMESH 2D is a mistake, sorry in advance (I believe it should only be added into the assets folder)

and incase I mess up I have the most recently saved file on my computer (prior to any of my commits) im still kinda new to github so sorry if Im missing anything. Just lmk and Ill reupload the last saved file